requirements.txt
python-dotenv
aiogram~=3.22.0
sqlalchemy~=2.0.43
aiosqlite~=0.21.0